Fix a race that may cause commit and push to fail
Defer the super.taskCompleted() call on DockerComputer until we told the
Docker slave node which build it has run last.  The DockerSlave needs
that link to the latest build in order to decide whether it should tag
and push the container.  However, if super.taskCompleted() is called too
early then the slave will already have been taken offline, the node will
have been deleted and slaveShutdown() will never be called.