feat(DockerCloud): option to use docker hostname as slave display name
Adds an option to the docker cloud to use the hostname of the docker
host as slave display name. If the hostname cannot be resolved the IP
address is used.

This is useful for setups with docker swarm. The display name of the
slave will then contain the information on which actual docker host
the job is running.

Limitation: The docker container must have a port mapping to
get the hostname of the slave.