Slightly re-work the PortUtils into a more fluent API, and re-name so that it's more obvious what action is taking place.
The old style had an API like this:
  portUtils.withRetries(60).withEveryRetryWaitFor(2, TimeUnit.SECONDS)

It's not immediately obvious that the final 'with' is, in fact different
to the previous one in that it actually does the work, since it doesn't
have a verb ending.

It's slightly odd that the SSH edition also tests with non-ssh first,
but that's what it used to do so leaving for now.

Signed-off-by: Nigel Magnay <nigel.magnay@gmail.com>