DockerCloud: Accompany "Bad template" message with actual stacktrace.
That message is confusing, and oftentimes, wrong - the problem may be not
with Docker template, but with something else, and it may be not even on
Docker side, but on plugin side. So, just let people to see what really
happens and where, to have better chance to diagnose it.

Implementation uses SLF4J feature available since 1.6:
http://www.slf4j.org/faq.html#paramException (Jenkins uses 1.7)

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>