Improve documentation on Retention Strategy/Availability drop down
The field for defining the retention strategy for a DockerTemplate
refers to the generic availability help. However, the values described
there only do have little in common with the values which the end user
can select. 
This commit provide a rough summary about the retention strategies and
tries to describe their behavior in a nutshell to the end user. The text
is newly associated to the field "Availability" in the configuration
dialog.