JENKINS-36080 enable to specify docker api version
Revert "Hide API version from users and nullify for existed installations."
This reverts commit 59abe7c1916c02247228b537be41d4ebd0478049.

Since our dependency docker-java is not yet docker api v1.24 compatible
we need a way to specify the docker api version to use. This was
functionality was removed a while ago. It is back now, as a more or less
workaround.

Note that "http://.." was removed from docker-java as a configuration option, it should now
be "tcp://..."