Introduction of Local Container Cap / Refactoring Global Container Cap
The existing Container Cap is global in that sense that it limits the
number of containers which may run globally on the docker server. If
running two Jenkins instances on the same docker server, this global
limit is hard to maintain properly, as both instances need to have the
same Container cap. Otherwise, the instance with the lower Container cap
will starve.
To allow to properly handle this scenario, a new configuration option
Local Container Cap is being introduced. This one does not consider all
containers running on the docker server, but only those, which were
created by the current Jenkins instance. Thus, this value limits the
number of containers which /this/ Jenkins instance may execute on the
docker server.
To ensure upward compatibility with existing configurations, the Local
Container Gap is defaulted with 0 (switches it off).
To allow proper distinction of the terms, the existing "Container Cap"
is refactored to be now called "Global Container Cap".