[FEATURE] Docker Watchdog Async Periodic Work - first version
As being discussed in
https://github.com/jenkinsci/docker-plugin/issues/401, there seems to be
issues with containers which are not cleaned up properly. Therefore, a
Watchdog-like periodic work instance is being introduced, which verifies
the consistency of the containers which are reported to be running on
the docker server (and were provisioned by the current Jenkins instance)
compared with the slaves attached to the current Jenkins instance.
Moreover, containers which have exited, but have not been removed yet,
are reported.

So far, only reports to the logs are being generated. After further
analysis on the issue has been performed, there is the intention to also
trigger clean up commands to the docker server, fixing container leaks.