[Fix] Help text on Container cap - Default value / value of zero
As seen in the discussion of pull request
https://github.com/jenkinsci/docker-plugin/pull/425, it turned out that
during the migration from a string parameter to a numerical parameter,
the semantics of "0" has been adjusted to mean "zero containers may be
provisioned". Unfortunately, the help text was forgotten to be adopted
properly.
This commit fixes the documentation to reflect the current state of
implementation.