Configure Docker hostname when using unix socks
When using Docker URL with unix-domains sockets we cannot access the
docker hostname from the host part of the URL. Even using the docker
inspection of the machine created dynamically would not expose the
right value accessible from the outside.

Allow to explicitly define the Docker hostname you want to use when
Jenkins master needs to connect to slave instances created.