BugFix: docker-java's Netty ListContainers depends on Jersey code.
The docker-java ListContainersCmdExec class used by the
NettyDockerCmdExecFactory calls upon Jersey functionality to implement
the filter.  This results in a classloading exception at runtime as
we're using Netty, and not Jersey.
This error is repeated in a number of xxxCmdExec classes, but this
commit only fixes it for the ListContainers command as that's the only
one we need fixed right now.