Bugfix: issue#615
The docker cloud "Test Connection" button causes socket resources to be
leaked.  The doTestConnection method in DockerAPI creates a new
DockerClient and then forgets about it, leaving it open.
This change adds a finally clause that closes the connection after the
test to ensure that it doesn't leak so easily.