Prevent invalid idleMinutes in retention strategy
The DockerOnceRetentionStrategy class changed the name of the field in
which it stored the "timeout" value.  This resulted in a lot of folks
finding it had become zero, and this caused the plugin to close
containers "due to inactivity" before they had a chance to become
active.
This change ensures that, in the event of an invalid value being
provided, the class will use the default value (10 minutes), and also
ensures that, from this point onwards, the value persisted is the value
read (previously there was a discrepancy).