Prevent over-provisioning
Jenkins can keep on asking for more executors even though it has enough
slaves promised to it that will satisfy demand once they come online.
We already keep track of the number of slaves that we are in the process
of creating but have yet to give to Jenkins, so if we include that count
when determining demand, we can ensure we do not provision more
containers than permitted.