Create means of caching DockerClient instances.
We need a means of caching, as connecting repeatedly is expensive.
We need to avoid resource leaking too.
We can't use a simple "discard after 10 minutes" timeout as a connection
can be used for a multi-hour operation - we need to ensure that we do
not throw anything way while it's "in use".

This cache tracks usage and only starts the timeout once an entry is no
longer in use.  Once it's been unused for too long, then it can be
discarded.
The downside is that it does require callers to always properly
"dispose" of the resource in order to keep track of things correctly.