Automatically temporarily disable malfunctioning templates/clouds.
DockerCloud and DockerTemplate now have a "Enabled" field so a user can
disable an entire cloud or template until they choose to re-enable it.
If a template fails to provision, this flag is cleared (on the template)
for 5 minutes.
If a cloud is unable to interrogate the docker container counts then the
flag is cleared (on the entire cloud) for 5 minutes.