Fix for issue #644
With Jenkins 2.107.2 and docker-plugin 1.1.4, we find that during node provisioning, if many nodes are simultaneously requested, it is possible to have the node's persistent filesystem yanked from under it while the provisioner is trying to set node properties causing an exception to raise. The exception will trigger a disableBySystem, which we don't want since the provisioning could just proceed normally. This is currently our best workaround.

See issue #644