Mitigate JENKINS-51203 and issue 644.
If there are no node properties to set then we avoid calling
setNodeProperties entirely.
If there are node properties to set then we try setting them up to 10
times before giving up.
And we no longer call Jenkins.addNode ourselves because the Jenkins core
NodeProvisioner will do that for us anyway (and do it safely).