Improved help for connect methods.
Prerequisites for all 3 methods are now bullet-pointed lists.
SSH now specifies that login details are needed as per ssh-slaves
plugin.
JNLP now specifies that slave.jar is needed.
JNLP's entrypoint argument help improved to make it clear that the
multi-line string needs to be expanded before being edited.
Example of custom command now mentions that it requires wget.
Also, moved details of the default setting to the bottom of the help
text.