Cloud/Template help improvements
Correction: Cloud and Template containerCap does *not* count containers that were not started by (this) Jenkins.
Correction: Help for template's environment variables field was mis-named as "environment" not "environmentString" so it did not appear.
Enhancement: Added extra advice for editing environment variables field as it's far too easy to accidentally corrupt whitespace.