FEATURE: add option to be able to use node name computed as container name.
To keep tracks builds, it is sometimes easier to quickly get the container build which is running directly from the computed nodename instead of looking into the docker page to get the container id.
It saves one step.
However, this feature is optional because most of time we don't care about the name container, but sometimes we do.