Prospective fix for #757 and JENKINS-59790.
Refactor DockerComputerConnector.beforeContainerStarted(...) and
afterContainerStarted(...) so it takes a node (that holds a containerId)
instead of a containerId.
Refactor DockerTransientNode constructors so the launcher can be
provided post-construction.
Refactor DockerTemplate.doProvision(...) so that the DockerTransientNode
is created without a launcher before we start the container and the
launcher is only set after the container is started.
Enhance DockerComputerJNLPConnector so that it adds the node to Jenkins
immediately before the container is created, ensuring that it'll be
known to Jenkins (just) before the JNLP process starts.