Improving uniqueness of templateID string
The instance capacity help text says that it is "The maximum number of containers, based on this template, that this provider is allowed to run in total.". With the current implementation, the instance capacity is based on the number of instances of an image, not of a template. This causes issues with different templates that use the same image. 
This change makes the templateID more unique, so a user can define different templates that use the same image as long as the labels and name are unique.