Do not test later Jenkins versions (yet)
Jenkins versions 2.357 and later have the instance identity module split
into a plugin.  This plugin needs the features of instance identity for
its tests.  Rather than test newer versions of Jenkins core, limit the
testing to 2.346.4 until the plugin is ready to require Java 11 and a
minimum Jenkins version of 2.361.4.