[JENKINS-70729] Stop supporting an empty cloud name
https://github.com/jenkinsci/jenkins/pull/7658 reworks cloud management
to use multiple pages instead of a single large page.  That rework also
requires that a cloud may not be named with an empty string.

The docker plugin specifically tested that a cloud was allowed to have an
empty string as its name, but performed no other operations with a cloud
that is named with an empty string.  Cloud named with an empty string may
have worked prior to Jenkins 2.403, but they would have been complicated
to manage, especially in a controller with multiple clouds defined.

A cloud named with an empty string is much more difficult to manage.
It is better to mention in the changelog that clouds are no longer
allowed to have an empty name rather than wrestling with all the ways
that an empty name will break cloud administration.

https://issues.jenkins.io/browse/JENKINS-70729 is the enhancement request
that implements the change.

https://issues.jenkins.io/browse/JENKINS-71825 is the bug report for
the automated test failure that started this investigation.