Convert empty or blank name to non-blank in readResolve
Jenkins 2.403 improves the cloud management user interface.  Each cloud is
now managed from a separate page.  That separate page requires a non-blank
name for the cloud.  Jenkins 2.402 and later allowed a blank name.

When a blank name is detected, convert it to a non-blank name based on
the hashCode of the DockerAPI and templates.