Allow randomizing cloud selection
When multiple Docker clouds are defined, the plugin will always create
agents from the first available cloud, not allowing to distribute the
load on all the clouds that are defined.

Possible solutions to this issue are:
- Docker Swarm Standalone, which is deprecated
- Docker Engine Swarm API, which is not supported by docker-plugin.

This changes adds a setting that enables randomizing the order in which
Docker clouds are tried when creating a new agent.