Update documetation
Docker swarm standalone is no longer developed.  The GitHub repository
has been archived.  Remove mentions of swarm mode, other than to note
that Docker Engine Swarm is not supported.

Link to the changelog on the plugins site since it is much more complete
than the changelog stored in the repository.